requirements.txt
# pyspark
# python-dotenv
apache-airflow-providers-google
# pyarrow
PyYAML
requests
python-dotenv